+++ src/shared/components/palette/utils/first-error.pipe.ts
@@ -6,24 +6,28 @@ import { ValidationErrors } from '@angular/forms';
 })
 export class FirstErrorPipe implements PipeTransform {
 
-  transform(value: ValidationErrors): string {
+  transform(value: ValidationErrors, args?: string): string {
     if (!value) {
       return '';
     }
 
-    let keys = Object.keys(value);
+    if (!args) {
+      args = 'field';
+    }
+
+    const keys = Object.keys(value);
 
     if (!keys.length) {
       return '';
     }
     if (keys[0] ===  'required') {
-      return 'This field is required';
+      return `${args} is required`;
     } else if (keys[0] ===  'pattern') {
-      return 'The data entered is not valid for this type of field.';
+      return `The data entered is not valid for ${args}`;
     } else if (keys[0] ===  'minlength') {
-      return 'Required minimum length';
+      return `${args} is below the minimum length`;
     } else if (keys[0] ===  'maxlength') {
-      return 'Exceeds maximum length';
+      return `${args} exceeds the maximum length`;
     }
     return value[keys[0]];
   }